Element-lokale Zeitmessung fuer Gewichtsberechung. 50% schneller als mit globalen Gewicht. Problem: Partikel fuehren zu einem Unterschied 2x in der Last zwischen einzelnen Procs
git-svn-id: svn+ssh://iagpc151.iag.uni-stuttgart.de/usr/local/svnroot/Boltzplatz/branches/PICASSO@881 f8cfc4a6-4e3a-43c3-8588-810d72ca601e

+++ src/loadbalance/loadbalance_vars.f90
@@ -43,9 +43,9 @@ REAL                                :: targetWeight                             
 REAL                                :: tCartMesh                                  ! time for CartMesh deposition
 REAL                                :: tTracking                                  ! time for relocation of particles
 REAL,ALLOCATABLE                    :: ElemTime(:)
-INTEGER,ALLOCATABLE                 :: nPartsPerElem(:)
-INTEGER,ALLOCATABLE                 :: nDeposPerElem(:)
-INTEGER,ALLOCATABLE                 :: nTracksPerElem(:)
+INTEGER(KIND=8),ALLOCATABLE         :: nPartsPerElem(:)
+INTEGER(KIND=8),ALLOCATABLE         :: nDeposPerElem(:)
+INTEGER(KIND=8),ALLOCATABLE         :: nTracksPerElem(:)
 
 
 END MODULE MOD_LoadBalance_Vars
